% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReliabilityAnalyzer.R
\name{ReliabilityAnalyzer}
\alias{ReliabilityAnalyzer}
\title{ReliabilityAnalyzer: Reliability and Survival Analysis Engine}
\description{
Performs reliability computations: distribution fitting, survival
estimation, hazard functions, and reliability metrics (MTTF, B10, B50).
Inherits from \link[iQualityR.core:IqrAnalyzerBase]{iQualityR.core::IqrAnalyzerBase} and uses the standardized results
container.

\strong{Core capabilities}:
\itemize{
\item Parametric distribution fitting (Weibull, exponential, lognormal, logistic)
\item Kaplan-Meier nonparametric survival estimation
\item Cox proportional hazards regression
\item Goodness-of-fit testing (Kolmogorov-Smirnov)
\item MTTF, B10, B50 life, and reliability calculations
}
}
\section{Super class}{
\code{\link[iQualityR.core:IqrAnalyzerBase]{iQualityR.core::IqrAnalyzerBase}} -> \code{ReliabilityAnalyzer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReliabilityAnalyzer-analyze}{\code{ReliabilityAnalyzer$analyze()}}
\item \href{#method-ReliabilityAnalyzer-clone}{\code{ReliabilityAnalyzer$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="get_results"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-get_results'><code>iQualityR.core::IqrAnalyzerBase$get_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="initialize"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-initialize'><code>iQualityR.core::IqrAnalyzerBase$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="reset"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-reset'><code>iQualityR.core::IqrAnalyzerBase$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="run"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-run'><code>iQualityR.core::IqrAnalyzerBase$run()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="set_datatable"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-set_datatable'><code>iQualityR.core::IqrAnalyzerBase$set_datatable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="set_diagnostic"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-set_diagnostic'><code>iQualityR.core::IqrAnalyzerBase$set_diagnostic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="set_raw_output"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-set_raw_output'><code>iQualityR.core::IqrAnalyzerBase$set_raw_output()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="set_statistic"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-set_statistic'><code>iQualityR.core::IqrAnalyzerBase$set_statistic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="setup"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-setup'><code>iQualityR.core::IqrAnalyzerBase$setup()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReliabilityAnalyzer-analyze"></a>}}
\if{latex}{\out{\hypertarget{method-ReliabilityAnalyzer-analyze}{}}}
\subsection{Method \code{analyze()}}{
Execute reliability analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReliabilityAnalyzer$analyze(data, plan)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame.}

\item{\code{plan}}{\link{ReliabilityPlan} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Structured list of computation results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReliabilityAnalyzer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReliabilityAnalyzer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReliabilityAnalyzer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
