% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReliabilityPlan.R
\name{ReliabilityPlan}
\alias{ReliabilityPlan}
\title{ReliabilityPlan: Reliability and Survival Analysis Plan Configurator}
\description{
Inherits from \link[iQualityR.core:IqrPlanBase]{iQualityR.core::IqrPlanBase} and stores all configuration parameters for a
reliability or survival analysis task.

\strong{Supported analysis types}:
\itemize{
\item \strong{Parametric reliability}: Weibull, exponential, lognormal, logistic
\item \strong{Nonparametric survival}: Kaplan-Meier estimation
\item \strong{Semiparametric survival}: Cox proportional hazards model
}
}
\section{Super class}{
\code{\link[iQualityR.core:IqrPlanBase]{iQualityR.core::IqrPlanBase}} -> \code{ReliabilityPlan}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{time_var}}{Character. Name of the time/lifetime variable.}

\item{\code{status_var}}{Character. Name of the status variable.}

\item{\code{censoring_type}}{Character. Censoring type.}

\item{\code{distribution}}{Character. Fitted distribution.}

\item{\code{method}}{Character. Analysis method.}

\item{\code{factors}}{Character vector. Covariates for the Cox model.}

\item{\code{stress_vars}}{Character vector. Stress variables for ALT.}

\item{\code{acceleration_model}}{Character. Acceleration model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReliabilityPlan-new}{\code{ReliabilityPlan$new()}}
\item \href{#method-ReliabilityPlan-validate_type}{\code{ReliabilityPlan$validate_type()}}
\item \href{#method-ReliabilityPlan-validate}{\code{ReliabilityPlan$validate()}}
\item \href{#method-ReliabilityPlan-is_censored}{\code{ReliabilityPlan$is_censored()}}
\item \href{#method-ReliabilityPlan-is_alt}{\code{ReliabilityPlan$is_alt()}}
\item \href{#method-ReliabilityPlan-clone}{\code{ReliabilityPlan$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlanBase" data-id="generate_protocol"><a href='../../iQualityR.core/html/IqrPlanBase.html#method-IqrPlanBase-generate_protocol'><code>iQualityR.core::IqrPlanBase$generate_protocol()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlanBase" data-id="set_criteria"><a href='../../iQualityR.core/html/IqrPlanBase.html#method-IqrPlanBase-set_criteria'><code>iQualityR.core::IqrPlanBase$set_criteria()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlanBase" data-id="set_meta"><a href='../../iQualityR.core/html/IqrPlanBase.html#method-IqrPlanBase-set_meta'><code>iQualityR.core::IqrPlanBase$set_meta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlanBase" data-id="to_list"><a href='../../iQualityR.core/html/IqrPlanBase.html#method-IqrPlanBase-to_list'><code>iQualityR.core::IqrPlanBase$to_list()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReliabilityPlan-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReliabilityPlan-new}{}}}
\subsection{Method \code{new()}}{
Initialize the reliability analysis plan.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReliabilityPlan$new(
  time_var,
  status_var = NULL,
  censoring_type = "right",
  distribution = "weibull",
  method = "parametric",
  factors = NULL,
  stress_vars = NULL,
  acceleration_model = "arrhenius",
  conf_level = 0.95,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_var}}{Character. Name of the time/lifetime variable.}

\item{\code{status_var}}{Character. Name of the status variable
(optional; defaults to no censoring).}

\item{\code{censoring_type}}{Character. Censoring type.}

\item{\code{distribution}}{Character. Fitted distribution.}

\item{\code{method}}{Character. Analysis method.}

\item{\code{factors}}{Character vector. Covariates for the Cox model.}

\item{\code{stress_vars}}{Character vector. Stress variables for ALT.}

\item{\code{acceleration_model}}{Character. Acceleration model.}

\item{\code{conf_level}}{Numeric. Confidence level, between 0 and 1.}

\item{\code{...}}{Additional arguments passed to \link[iQualityR.core:IqrPlanBase]{iQualityR.core::IqrPlanBase}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReliabilityPlan-validate_type"></a>}}
\if{latex}{\out{\hypertarget{method-ReliabilityPlan-validate_type}{}}}
\subsection{Method \code{validate_type()}}{
Validate enum values (called by \code{initialize()}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReliabilityPlan$validate_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReliabilityPlan-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ReliabilityPlan-validate}{}}}
\subsection{Method \code{validate()}}{
Validate data and configuration compatibility.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReliabilityPlan$validate(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReliabilityPlan-is_censored"></a>}}
\if{latex}{\out{\hypertarget{method-ReliabilityPlan-is_censored}{}}}
\subsection{Method \code{is_censored()}}{
Whether the data is censored.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReliabilityPlan$is_censored()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReliabilityPlan-is_alt"></a>}}
\if{latex}{\out{\hypertarget{method-ReliabilityPlan-is_alt}{}}}
\subsection{Method \code{is_alt()}}{
Whether this is an accelerated life test plan.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReliabilityPlan$is_alt()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReliabilityPlan-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReliabilityPlan-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReliabilityPlan$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
